+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
  - new action for clustering nuvlaboxes
+ - new action for cleaning up leftover NuvlaBox clusters
 
 ## [2.14.2] - 2021-04-15