+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+  - nuvlabox_decommission - remove online attribute when nuvlabox is
+    decommissioned
+
 ## [2.14.2] - 2021-04-15
 
 ### Changed